+++ test/noahtheduke/splint/re_frame_test.clj
@@ -38,7 +38,7 @@
 
 (defexpect ^:integration re-frame-test
   (let [re-frame (gl/procure "https://github.com/day8/re-frame.git" 'day8/re-frame "v1.3.0")
-        results (run-impl [re-frame]
+        results (run-impl [{:path re-frame}]
                           (-> all-enabled-config
                               (assoc :silent true)
                               (assoc :parallel false)